+++ changelog.md
@@ -10,6 +10,7 @@
 - Fixed export of keyframes with constant interpolation 
 - Clean up View Panel. Now properties are hidden for non-Auth3D objects 
 - Fixed reading of type=0 channels for visibility 
+- Fixed hrc export crashing on armatures with no animation 
 
 
 ## 1.0.24